refactor(browser): extract browser/audition/remote UI state into browser_state
